+++ frontend/document_collaboration_system/src/accordion.js
@@ -6,7 +6,7 @@ import AccordionSummary from "@material-ui/core/AccordionSummary";
 import Typography from "@material-ui/core/Typography";
 import ExpandMoreIcon from "@material-ui/icons/ExpandMore";
 import { Button, TextField } from "@material-ui/core";
-import api from "./service";
+import api from "./services/APIService";
 import { useHistory } from "react-router-dom";
 import { AppContext } from "./";
 import "./css/base.css";
