+++ frontend/document_collaboration_system/src/chat.js
@@ -5,7 +5,7 @@ import ListItemIcon from "@material-ui/core/ListItemIcon";
 import MessageIcon from "@material-ui/icons/Message";
 import ListItemText from "@material-ui/core/ListItemText";
 import List from "@material-ui/core/List";
-import api from "./service";
+import api from "./services/APIService";
 import { AppContext } from "./";
 import _ from "lodash";
 import { useHistory } from "react-router-dom";
