+++ frontend/document_collaboration_system/src/commentWidgets.js
@@ -3,7 +3,7 @@ import Button from "@material-ui/core/Button";
 import EditCommentDialog from "./editCommentsDialog";
 import CommentsDialog from "./commentsDialog";
 import "./css/base.css";
-import api from "./service";
+import api from "./services/APIService";
 
 export function CommentButton({ document }) {
   const [openCommentsList, setOpenCommentsList] = useState(false);
