+++ frontend/document_collaboration_system/src/editCommentsDialog.js
@@ -7,7 +7,7 @@ import DialogContentText from "@material-ui/core/DialogContentText";
 import TextField from "@material-ui/core/TextField";
 import DialogActions from "@material-ui/core/DialogActions";
 import Button from "@material-ui/core/Button";
-import api from "./service";
+import api from "./services/APIService";
 
 export default function EditCommentDialog({
   openedComment,
