+++ frontend/document_collaboration_system/src/login.js
@@ -3,7 +3,7 @@ import React, { useCallback, useContext, useState } from "react";
 import "./css/login.css";
 import { Link, useHistory } from "react-router-dom";
 import { AppContext } from "./";
-import api from "./service";
+import api from "./services/APIService";
 
 function Login() {
   const [textField, setText] = useState("");
