+++ frontend/document_collaboration_system/src/messageButton.js
@@ -3,7 +3,8 @@ import Button from "@material-ui/core/Button";
 import Menu from "@material-ui/core/Menu";
 import MenuItem from "@material-ui/core/MenuItem";
 import SendMessageDialog from "./messageDialog";
-import api from "./service";
+import api from "./services/APIService";
+
 
 export function MessageButton({ document }) {
   const [anchorEl, setAnchorEl] = useState(null);
