+++ frontend/document_collaboration_system/src/messageDialog.js
@@ -11,7 +11,7 @@ import MenuItem from "@material-ui/core/MenuItem";
 import { TextField } from "@material-ui/core";
 import DialogActions from "@material-ui/core/DialogActions";
 import Button from "@material-ui/core/Button";
-import api from "./service";
+import api from "./services/APIService";
 
 export default function SendMessageDialog({
   openDialog,
