+++ frontend/document_collaboration_system/src/placeCommentDialog.js
@@ -7,7 +7,7 @@ import DialogContent from "@material-ui/core/DialogContent";
 import DialogContentText from "@material-ui/core/DialogContentText";
 import DialogTitle from "@material-ui/core/DialogTitle";
 import { AppContext } from "./";
-import api from "./service";
+import api from "./services/APIService";
 
 export default function FormDialog({
   openModalWindow,
